@@ -1,13 +1,15 @@
+import 'package:app/data/local/theme_data_source.dart';
+import 'package:app/data/provider/theme_data_source_provider.dart';
+import 'package:app/data/repository/theme_repository.dart';
 import 'package:flutter/material.dart';
-
-import '../local/theme_data_source.dart';
-import 'theme_repository.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 class ThemeRepositoryImpl implements ThemeRepository {
-  ThemeRepositoryImpl({required ThemeDataSource dataSource})
-      : _dataSource = dataSource;
+  ThemeRepositoryImpl(this._reader);
 
-  final ThemeDataSource _dataSource;
+  final Reader _reader;
+
+  late final ThemeDataSource _dataSource = _reader(themeDataSourceProvider);
 
   @override
   Future<ThemeMode?> loadThemeMode() {
