@@ -1,14 +1,13 @@
+import 'package:app/constants.dart';
+import 'package:app/ui/app_theme.dart';
+import 'package:app/ui/detail/detail_page.dart';
+import 'package:app/ui/home/home_page.dart';
+import 'package:app/ui/signIn/sign_in_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:get/get.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-import 'constants.dart';
-import 'ui/app_theme.dart';
-import 'ui/detail/detail_page.dart';
-import 'ui/home/home_page.dart';
-import 'ui/signIn/sign_in_page.dart';
 
 class App extends HookWidget {
   @override
@@ -16,8 +15,10 @@ class App extends HookWidget {
     final appTheme = context.read(appThemeNotifierProvider);
     final setting =
         useProvider(appThemeNotifierProvider.select((value) => value.setting));
-    useFuture(useMemoized(appTheme.themeMode, [setting]),
-        initialData: ThemeMode.light);
+    useFuture(
+      useMemoized(appTheme.themeMode, [setting]),
+      initialData: ThemeMode.light,
+    );
     return GetMaterialApp(
       title: 'Flutter Architecture Blueprints',
       theme: lightTheme,
