@@ -1,8 +1,8 @@
+import 'package:app/data/model/source.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'source.dart';
-
 part 'article.freezed.dart';
+
 part 'article.g.dart';
 
 @freezed
