@@ -1,6 +1,5 @@
+import 'package:app/data/local/app_shared_preferences.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-import '../local/app_shared_preferences.dart';
 
 final prefsProvider =
     Provider<AppSharedPreferences>((ref) => AppSharedPreferences());
