@@ -1,8 +1,6 @@
+import 'package:app/data/repository/auth_repository.dart';
+import 'package:app/data/repository/auth_repository_impl.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../repository/auth_repository.dart';
-import '../repository/auth_repository_impl.dart';
-import 'auth_data_source_provider.dart';
-
-final authRepositoryProvider = Provider<AuthRepository>(
-    (ref) => AuthRepositoryImpl(ref.read(authDataSourceProvider)));
+final authRepositoryProvider =
+    Provider<AuthRepository>((ref) => AuthRepositoryImpl(ref.read));
