@@ -1,7 +1,5 @@
+import 'package:app/data/local/theme_data_source_impl.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../local/theme_data_source_impl.dart';
-import 'app_shared_preferences_provider.dart';
-
 final themeDataSourceProvider =
-    Provider((ref) => ThemeDataSourceImpl(ref.read(prefsProvider)));
+    Provider((ref) => ThemeDataSourceImpl(ref.read));
