@@ -1,8 +1,6 @@
+import 'package:app/data/repository/theme_repository.dart';
+import 'package:app/data/repository/theme_repository_impl.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../repository/theme_repository.dart';
-import '../repository/theme_repository_impl.dart';
-import 'theme_data_source_provider.dart';
-
-final themeRepositoryProvider = Provider<ThemeRepository>((ref) =>
-    ThemeRepositoryImpl(dataSource: ref.read(themeDataSourceProvider)));
+final themeRepositoryProvider =
+    Provider<ThemeRepository>((ref) => ThemeRepositoryImpl(ref.read));
