@@ -1,6 +1,5 @@
-import '../model/news.dart';
+import 'package:app/data/model/news.dart';
 
-// ignore: one_member_abstracts
 abstract class NewsDataSource {
   Future<News> getNews();
 }
