@@ -1,6 +1,5 @@
+import 'package:app/data/model/result.dart';
 import 'package:firebase_auth/firebase_auth.dart' as firebase;
-
-import '../model/result.dart';
 
 abstract class AuthRepository {
   Future<Result<firebase.User?>> signIn();
