@@ -1,7 +1,6 @@
-import '../model/news.dart';
-import '../model/result.dart';
+import 'package:app/data/model/news.dart';
+import 'package:app/data/model/result.dart';
 
-// ignore: one_member_abstracts
 abstract class NewsRepository {
   Future<Result<News>> getNews();
 }
