@@ -1,10 +1,9 @@
+import 'package:app/constants.dart';
+import 'package:app/data/model/article.dart';
+import 'package:app/ui/component/image.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-
-import '../../constants.dart';
-import '../../data/model/article.dart';
-import 'image.dart';
 
 class ArticleItem extends StatelessWidget {
   const ArticleItem(this._article);
