@@ -1,9 +1,8 @@
+import 'package:app/ui/component/loading.dart';
+import 'package:app/ui/loading_state_view_model.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-
-import '../loading_state_view_model.dart';
-import 'loading.dart';
 
 class ContainerWithLoading extends StatelessWidget {
   const ContainerWithLoading({required Widget child}) : _child = child;
