@@ -2,8 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:get/get.dart';
 
-void showSimpleDialog(BuildContext context,
-    {String? title, String? text, VoidCallback? onPressed}) {
+void showSimpleDialog(
+  BuildContext context, {
+  String? title,
+  String? text,
+  VoidCallback? onPressed,
+}) {
   showDialog<void>(
       context: context,
       builder: (context) {
