@@ -1,27 +1,31 @@
+import 'package:app/constants.dart';
+import 'package:app/ui/app_theme.dart';
+import 'package:app/ui/component/article_item.dart';
+import 'package:app/ui/component/container_with_loading.dart';
+import 'package:app/ui/component/image.dart';
+import 'package:app/ui/home/home_view_model.dart';
+import 'package:app/ui/hook/use_l10n.dart';
+import 'package:app/ui/hook/use_theme.dart';
+import 'package:app/ui/loading_state_view_model.dart';
+import 'package:app/ui/user_view_model.dart';
+import 'package:app/util/error_snackbar.dart';
+import 'package:app/util/ext/async_snapshot.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:get/get.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../constants.dart';
-import '../../util/error_snackbar.dart';
-import '../../util/ext/async_snapshot.dart';
-import '../app_theme.dart';
-import '../component/article_item.dart';
-import '../component/container_with_loading.dart';
-import '../component/image.dart';
-import '../loading_state_view_model.dart';
-import '../user_view_model.dart';
-import 'home_view_model.dart';
-
-class HomePage extends StatelessWidget {
+class HomePage extends HookWidget {
   @override
   Widget build(BuildContext context) {
+    final theme = useTheme();
+    final l10n = useL10n();
     return Scaffold(
         appBar: AppBar(
-          title: Text(L10n.of(context)!.home,
-              style: Theme.of(context).textTheme.headline1),
+          title: Text(
+            l10n.home,
+            style: theme.textTheme.headline1,
+          ),
           actions: [
             // action button
             IconButton(
@@ -30,8 +34,10 @@ class HomePage extends StatelessWidget {
                   .read(appThemeNotifierProvider)
                   .toggle()
                   .catchError((error) {
-                showErrorSnackbar(L10n.of(context)!.error,
-                    L10n.of(context)!.failedSwitchTheme);
+                showErrorSnackbar(
+                  l10n.error,
+                  l10n.failedSwitchTheme,
+                );
               }),
             ),
             IconButton(
